Add JWT authentication and authorization to AuctionService
Integrate JWT-based authentication using IdentityService and enforce authorization on critical endpoints. Updated business logic to verify seller identity for modifications and deletions. Added configurations for token handling and included the required dependencies.

+++ Carsties/src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -42,6 +43,7 @@ public class AuctionsController: ControllerBase
 
     }
 
+    
     [HttpGet("{id:guid}")]
     public async Task<ActionResult<AuctionDto>> GetAuctionById(Guid id)
     {
@@ -55,13 +57,15 @@ public class AuctionsController: ControllerBase
         return _mapper.Map<AuctionDto>(auction);
     }
     
+    
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
     {
         var auction = _mapper.Map<Auction>(auctionDto);
         // TODO: add current user as seller
-        auction.Seller = "test";
-        
+        if (User.Identity != null) auction.Seller = User.Identity.Name;
+
         _context.Auctions.Add(auction);
         
         var newAction = _mapper.Map<AuctionDto>(auction);
@@ -80,6 +84,7 @@ public class AuctionsController: ControllerBase
             new {auction.Id}, newAction);
     }
     
+    [Authorize]
     [HttpPut("{id:guid}")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
     {
@@ -91,7 +96,7 @@ public class AuctionsController: ControllerBase
             return NotFound();
         }
         
-        // TODO: check seller == username
+        if (User.Identity != null && auction.Seller != User.Identity.Name) return Forbid();
         
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -109,6 +114,7 @@ public class AuctionsController: ControllerBase
         return Ok();
     }
 
+    [Authorize]
     [HttpDelete("{id:guid}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -116,7 +122,7 @@ public class AuctionsController: ControllerBase
 
         if (auction == null) return NotFound();
         
-        // TODO: check seller == username
+        if(User.Identity != null && auction.Seller != User.Identity.Name) return Forbid();
 
         _context.Auctions.Remove(auction);
